refactor: rename list-access.ts utils to containers.ts
These functions are now used for all container logic.
